--- src/main/java/org/example/battleunits/CanAttack.java
@@ -1,5 +1,0 @@
-package org.example.battleunits;
-
-public interface CanAttack {
-    int getAttack();
-}
