--- src/main/java/org/example/battleunits/HasDefence.java
@@ -1,5 +1,0 @@
-package org.example.battleunits;
-
-public interface HasDefence {
-    int getDefence();
-}
