--- src/main/java/org/example/battleunits/HasVampirism.java
@@ -1,6 +1,0 @@
-package org.example.battleunits;
-
-public interface HasVampirism extends CanAttack {
-    int getVampirism();
-
-}
