+++ src/main/java/org/example/battleunits/characteristic/Attack.java
@@ -1,0 +1,5 @@
+package org.example.battleunits.characteristic;
+
+public interface Attack {
+    int getAttack();
+}
