+++ src/main/java/org/example/battleunits/characteristic/Defence.java
@@ -1,0 +1,5 @@
+package org.example.battleunits.characteristic;
+
+public interface Defence {
+    int getDefence();
+}
