+++ src/main/java/org/example/battleunits/characteristic/Vampirism.java
@@ -1,0 +1,6 @@
+package org.example.battleunits.characteristic;
+
+public interface Vampirism {
+    int getVampirism();
+
+}
